maxx_ai/tf/transfer.py
import onnx

model = onnx.load('.models/classifier_model.onnx')

import tensorflow as tf
import tf2onnx

# Convert ONNX model to TensorFlow model
tf_model = tf2onnx.from_onnx(model)

# Modify the TensorFlow model as needed
# ...

# Compile the model
tf_model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])

import cv2
import numpy as np

# Load images and labels
train_images = []
train_labels = []

for image_path, label in zip(image_paths, labels):
    image = cv2.imread(image_path)
    image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)  # Convert to RGB if needed
    train_images.append(image)
    train_labels.append(label)

train_images = np.array(train_images)
train_labels = np.array(train_labels)

# Train the model
tf_model.fit(train_images, train_labels, batch_size=32, epochs=10)

# Save the retrained model as a .h5 file
tf_model.save('path/to/retrained_model.h5')