requirement2.txt
absl-py==2.0.0
asgiref==3.7.2
astunparse==1.6.3
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Django==4.2.6
dm-tree==0.1.8
filelock==3.13.0
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.0
h5py==3.10.0
humanfriendly==10.0
idna==3.4
imageio==2.31.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.4.5
labeling==0.1.13
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mmdnn==0.3.1
mpmath==1.3.0
mysqlclient==2.2.0
networkx==3.2
nudenet==3.0.8
NudeNetClassifier @ git+https://github.com/platelminto/NudeNet@05ad8a29b4a2474e6cc4f7f76c76e77e01b08744
numpy==1.24.3
oauthlib==3.2.2
onnx==1.15.0
onnxconverter-common==1.14.0
onnxruntime==1.15.0
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
Pillow==9.5.0
progressbar==2.5
protobuf==3.20.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.15.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.20.1
termcolor==2.3.0
tf2onnx==1.15.1
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.15.0